vista de reserva y edicion de las mismas desde propietario

+++ QuinchApp/src/main/java/com/QuinchApp/Servicios/ReservaServicio.java
@@ -45,8 +45,8 @@ public class ReservaServicio {
         reserva.setConfirmada(activo);
         Date fecha = new SimpleDateFormat("yyyy-MM-dd").parse(fechaDelEvento);
         List<Reserva> reservas = propiedadReserva.getReservas();
-        for(Reserva r : reservas) {
-            if(r.getFechaDelEvento().equals(fecha)) {
+        for (Reserva r : reservas) {
+            if (r.getFechaDelEvento().equals(fecha)) {
                 throw new Exception("error");
             }
         }
@@ -82,14 +82,31 @@ public class ReservaServicio {
         return reserva;
     }
 
+    @Transactional
+    public List<Reserva> listarResevas(String palabraClave) {
+        if (palabraClave != null) {
+            List<Reserva> reservas = reservaRepositorio.findAll(palabraClave);
+            return reservas;
+        } else {
+            List<Reserva> reservas = reservaRepositorio.findAll();
+            return reservas;
+        }
+    }
+
     public Reserva getOne(Integer id) {
         return reservaRepositorio.getOne(id);
     }
 
     @Transactional
-    public void borrar(Integer id) {
-        reservaRepositorio.deleteById(id);
-    }
+    public void borrar(Integer id) throws Exception {
+        Optional<Reserva> reservaOptional = reservaRepositorio.findById(id);
+        if (reservaOptional.isPresent()) {
+            reservaRepositorio.deleteById(id);
+        } else {
+            throw new Exception("La reserva con ID " + id + " no existe.");
+        }
+   }
+
 
     public Reserva bajaAlta(Integer id) {
         Optional<Reserva> reservaOptinal = reservaRepositorio.findById(id);
